+++ src/FormCompletionAssistant/index.ts
@@ -3,5 +3,3 @@ export * from "./FormFieldCompletionAssistant";
 export * from "./FormSectionCompletionAssistant";
 export * from "./IntegerFieldCompletionAssistant";
 export * from "./DateFieldCompletionAssistant";
-
-export * from "./types";
