+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./Assertion";
 export * from "./FormCompletionAssistant";
+export * from "./types";
